Modularized home default screen to bulletin
